Menus: drop Planning + Git, rename Graph to Query
- Planning category folded into Analysis. The three planning tools
  (steelman, taboo, murphyjitsu) move to analysis/ and take category:
  'analysis'. Tool IDs keep their 'planning.' prefix for stable
  references. Planning submenu removed from the editor right-click and
  the title-bar tool categories.
- Git menu removed. Only entry was a wired Commit All and two disabled
  Push/Pull placeholders — not useful yet. The underlying
  api.git.commit IPC stays for future re-wiring.
- Graph menu renamed to Query to match what it actually is: a
  query-centric surface (SPARQL + SQL, stock + saved, rebuild index).
