+++ src/renderer/lib/components/Editor.svelte
@@ -116,7 +116,6 @@
   }: Props = $props();
 
   const analysisTools = getToolInfosByCategory('analysis');
-  const planningTools = getToolInfosByCategory('planning');
   const learningTools = getToolInfosByCategory('learning');
 
   let editorContainer: HTMLDivElement;
@@ -779,7 +778,7 @@
         <button onclick={() => handleMenuAction(() => onInsertQueryList?.())}>Link List for Tag...</button>
       </div>
     </div>
-    {#if onToolInvoke && (analysisTools.length > 0 || planningTools.length > 0 || learningTools.length > 0)}
+    {#if onToolInvoke && (analysisTools.length > 0 || learningTools.length > 0)}
       <div class="separator"></div>
       {#if learningTools.length > 0}
         <div class="submenu-item" onmouseenter={adjustSubmenu}>
@@ -796,16 +795,6 @@
           <span class="submenu-trigger">Analysis &#x25B8;</span>
           <div class="submenu">
             {#each analysisTools as tool}
-              <button onclick={() => handleMenuAction(() => onToolInvoke?.(tool.id))}>{tool.name}</button>
-            {/each}
-          </div>
-        </div>
-      {/if}
-      {#if planningTools.length > 0}
-        <div class="submenu-item" onmouseenter={adjustSubmenu}>
-          <span class="submenu-trigger">Planning &#x25B8;</span>
-          <div class="submenu">
-            {#each planningTools as tool}
               <button onclick={() => handleMenuAction(() => onToolInvoke?.(tool.id))}>{tool.name}</button>
             {/each}
           </div>
