+++ src/shared/tools/registry.ts
@@ -45,6 +45,5 @@ function toInfo(tool: ThinkingToolDef): ThinkingToolInfo {
 export const CATEGORIES: { id: ToolCategory; label: string }[] = [
   { id: 'learning', label: 'Learning' },
   { id: 'analysis', label: 'Analysis' },
-  { id: 'planning', label: 'Planning' },
   { id: 'research', label: 'Research' },
 ];
